refactor(cymbal): rename posthog_utils->analytics, group rule modules
- modes/processing/posthog_utils -> analytics (mirrors core::analytics).
- Group the four rule kinds under modes/processing/rules/: assignment,
  suppression, spike (moved from the processing root) and grouping (moved
  out of fingerprinting/). TeamManager caches them together.
