+++ rust/cymbal/src/modes/processing/rules/mod.rs
@@ -1,0 +1,8 @@
+//! Per-team rule kinds the processing pipeline evaluates: assignment,
+//! suppression, spike-alerting, and grouping (custom fingerprints). Cached
+//! together by `crate::modes::processing::teams::TeamManager`.
+
+pub mod assignment;
+pub mod grouping;
+pub mod spike;
+pub mod suppression;
