+++ rust/cymbal/src/modes/processing/rules/spike.rs
