+++ rust/cymbal/src/modes/processing/stages/linking/issue.rs
@@ -8,14 +8,14 @@ use uuid::Uuid;
 
 use crate::{
     app_context::AppContext,
-    assignment_rules::{try_assignment_rules, Assignment},
+    modes::processing::rules::assignment::{try_assignment_rules, Assignment},
     error::UnhandledError,
     issue_resolution::{
         send_fingerprint_issue_state, send_issue_created_alert, send_issue_reopened_alert,
         send_new_fingerprint_event, Issue, IssueFingerprintOverride,
     },
     metric_consts::{ISSUE_CREATED, ISSUE_LINKER_OPERATOR},
-    posthog_utils::capture_issue_created,
+    analytics::capture_issue_created,
     stages::{linking::LinkingStage, pipeline::HandledError},
     teams::TeamManager,
     types::{
